\name{carryoverratebasic}
\alias{carryoverratebasic}
\title{carryoverratebasic}
\usage{
carryoverratebasic(dependent, marketsize, media1, media2, marketvar, data, mode)
}
\description{
carryoverratebasic
}
\examples{
carryoverratebasic(dependent, marketsize, media1, media2, marketvar, data, mode)
}
