\name{equitargetlevels}
\alias{equitargetlevels}
\title{equitargetlevels}
\usage{
equitargetlevels(startpoint, endpoint,numberoflevels,group1,group2,startMonth,endMonth,x_range, y_range,bilconversion)
}
\description{
equitargetlevels
}
\examples{
equitargetlevels(startpoint, endpoint,numberoflevels,group1,group2,startMonth,endMonth,x_range, y_range,bilconversion)
}
