\name{make_objfunc_gradient}
\alias{make_objfunc_gradient}
\title{make_objfunc_gradient}
\usage{
make_objfunc_gradient(dffincoeff,atllist,grpgroup,pricepergrp,digitalra,atlra,digitalnu)
}
\description{
make_objfunc_gradient
}
\examples{
make_objfunc_gradient(dffincoeff,atllist,grpgroup,pricepergrp,digitalra,atlra,digitalnu)
}
